requirements.txt
imageio==2.3.0
matplotlib==2.2.2
numpy==1.14.3
opencv-python==4.0.0.21
pandas==0.23.0
Pillow==5.1.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
torch==1.0.0
torchvision==0.2.1
tqdm==4.26.0